// src/interfaces/user.ts
export type Profile = {
  _id: string;
  phone: string;
  username?: string;
  first_name?: string;
  last_name?: string;
  birthday?: string;
  email: string;
  gender?: string;
  role?: string;
  status?: string;
  is_verified?: string;
  avatar_url?: string;
  description?: string;
};